01_10_orm_basics/requirements.txt
sqlalchemy==1.4.39
alembic==1.7.6
fastapi==0.73.0
psycopg2-binary==2.9.3
uvicorn==0.16.0
requests==2.28.2
pydantic==1.10.7
python-dotenv==1.0.0
pylint==2.17.1